+++ src/App.css
@@ -19,6 +19,7 @@
 body,
 html {
   background-color: var(--bg);
+  margin-bottom: 0;
 }
 button {
   border: none;
@@ -65,6 +66,17 @@ a {
   right: 0%;
   transform: translate(-50%, -50%);
   cursor: pointer;
+}
+.logo {
+  background: linear-gradient(
+    var(--linear-title-bg-left),
+    var(--linear-title-bg-right)
+  );
+  -webkit-background-clip: text;
+  background-clip: text;
+  color: transparent;
+  font-weight: bold;
+  font-size: 1.5rem;
 }
 
 @media only screen and (max-width: 390px) {
