+++ src/App.tsx
@@ -2,6 +2,7 @@ import "./App.css";
 import About from "./component/About/About";
 import BasicInfo from "./component/BasicInfo/BasicInfo";
 import Contact from "./component/Contact/Contact";
+import Footer from "./component/Footer/Footer";
 import Navbar from "./component/Navbar/Navbar";
 import Project from "./component/Project/Project";
 
@@ -13,6 +14,7 @@ function App() {
       <Project/>
       <About/>
       <Contact/>
+      <Footer/>
     </>
   );
 }
