+++ src/component/BasicInfo/Information.tsx
@@ -1,10 +1,7 @@
-import linkedIn from "../../assets/LinkedIn.svg";
-import github from "../../assets/GitHub.svg";
-import medium from "../../assets/Medium.svg";
 import Button from "../Button/Button";
 import ResumeDownload from "../ResumeDownload/ResumeDownload";
 import styles from "./BasicInfo.module.css";
-import ImageLink from "../ImageLink/ImageLink";
+import SocialLink from "./SocialLink";
 
 const Information = () => {
   return (
@@ -12,14 +9,7 @@ const Information = () => {
       <p className={styles.greeting}>Hi I am</p>
       <p>Inshiya Ravat</p>
       <h1 className={styles.designation}>Software Engineer</h1>
-      <div className={styles.socialLinks}>
-        <ImageLink
-          link="https://www.linkedin.com/in/inshiya-ravat-4b1372226/"
-          image={linkedIn}
-        />
-        <ImageLink link="https://github.com/InshiyaRavat" image={github} />
-        <ImageLink link="https://medium.com/@insravat" image={medium} />
-      </div>
+      <SocialLink/>
       <div className={styles.actionContainer}>
         <Button>
           <a href="#contact">Contact Me</a>
