+++ src/component/Footer/Footer.module.css
@@ -1,0 +1,22 @@
+.footer{
+    text-align: center;
+    background-color: var(--dark-grey);
+    margin-bottom: 0;
+    padding: var(--gap-large);
+}
+.listItems{
+    display: flex;
+    justify-content: center;
+    gap: var(--gap-small);
+}
+.social{
+    display: flex;
+    justify-content: center;
+}
+.mail{
+    color: var(--primary-orange);
+    text-decoration: underline;
+}
+.copyright{
+    margin-bottom: 0;
+}
