+++ src/index.css
@@ -1,3 +1,4 @@
 #root {
   width: 100%;
+  margin-bottom: 0;
 }
